feat(spinner): make progeny_core maintain the process registry as a temporary SQLite db
Previously, the running processes were tracked as a simple list of tuples in the ProcessRegistry class. Now, each Prodigy instance's identifier, session_name, port, pid and expiry_timestamp are stored in a SQLite DB created at the time of ProcessRegistry's instantiation, and removed at the end of its lifetime. This means we can start progenitor instances in separate processes and tap into the same shared registry. This should make it easier to use progeny_core in a web server with multiple workers.